Created a cloud backup of spiralpod holder
This idea comes from my need of having a convenient way to store Nespresso pods.
Goal: make a spiral/tree pod holder that is dependent on the number of pods you'd like to store.
Right now this is just the base, of which I've made 4 different feet type because I'm not sure what I like best. Next step is generating the algorithm for a spiral that doesn't intersect other pods.